requirements.txt
bpython==0.19
celery==4.4.2
coreapi==2.3.3
Django==2.2.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-filter==2.1.0
gunicorn==20.0.4
pyzabbix==0.7.5
PyMySQL==0.9.3
redis==3.5.2
